Receive xrandr & dpms exts correctly in tests
This makes tests pass that were failing with null reference errors.

`XClient.require` was passing the callback function to `requireExt`,
which called it with parameters `null, ext` on success. The callback
itself seemed to assume it will get only one parameter, which is either
the extension on an error, hence why it tried to call methods on a
`null`.